Validate deposit and withdrawal amounts.
Validation and model state set in controller. This is not ideal.

+++ Api/Request/OpenAccountRequest.cs
@@ -2,9 +2,9 @@ using Api.Data;
 
 namespace Api.Request;
 
-public class OpenAccountRequest(int customerId, AccountType accountType, double initialDeposit)
+public class OpenAccountRequest(int customerId, AccountType accountType, decimal initialDeposit)
 {
     public int CustomerId { get; } = customerId;
     public AccountType AccountType { get; } = accountType;
-    public double InitialDeposit { get; } = initialDeposit;
+    public decimal InitialDeposit { get; } = initialDeposit;
 }
